feat: migrate to a single command interface
