+++ docker/web/development/crontab
@@ -16,7 +16,5 @@
 0       2   * * * { . /code/config/bash.env && /usr/bin/php /code/web database:prospective-members:delete-expired; } > /data/logs/cron-delete-prospective.log 2>&1
 50      *   * * * { . /code/config/bash.env && /usr/bin/php /code/web database:mailman:fetch; } > /data/logs/cron-mailman-fetch.log 2>&1
 55      *   * * * { . /code/config/bash.env && /usr/bin/php /code/web database:listmonk:fetch; } > /data/logs/cron-listmonk-fetch.log 2>&1
-2       *   * * * { . /code/config/bash.env && /usr/bin/php /code/web database:local:syncmembership -f -vv; } > /data/logs/cron-local-syncmembership.log 2>&1
-5       *   * * * { . /code/config/bash.env && /usr/bin/php /code/web database:mailman:syncmembership -f -vv; } > /data/logs/cron-mailman-syncmembership.log 2>&1
-10      *   * * * { . /code/config/bash.env && /usr/bin/php /code/web database:listmonk:syncmembership -f -vv; } > /data/logs/cron-listmonk-syncmembership.log 2>&1
+5       *   * * * { . /code/config/bash.env && /usr/bin/php /code/web database:mailinglist:sync-membership -f -vv; } > /data/logs/cron-sync-membership.log 2>&1
 40      2   * * * { . /code/config/bash.env && /usr/bin/php /code/web database:mailinglist:maintenance -f -vv; } > /data/logs/cron-mailinglist-maintenance.log 2>&1
