+++ module/Database/src/Module.php
@@ -10,18 +10,14 @@ use Database\Command\Factory\DeleteExpiredMembersCommandFactory;
 use Database\Command\Factory\DeleteExpiredProspectiveMembersCommandFactory;
 use Database\Command\Factory\GenerateAuthenticationKeysCommandFactory;
 use Database\Command\Factory\ListmonkFetchListsCommandFactory;
-use Database\Command\Factory\ListmonkSyncMembershipCommandFactory;
 use Database\Command\Factory\MailingListMaintenanceCommandFactory;
-use Database\Command\Factory\MailingListSyncLocalMembershipCommandFactory;
+use Database\Command\Factory\MailingListSyncCommandFactory;
 use Database\Command\Factory\MailmanFetchListsCommandFactory;
-use Database\Command\Factory\MailmanSyncMembershipCommandFactory;
 use Database\Command\GenerateAuthenticationKeysCommand;
 use Database\Command\ListmonkFetchListsCommand;
-use Database\Command\ListmonkSyncMembershipCommand;
 use Database\Command\MailingListMaintenanceCommand;
-use Database\Command\MailingListSyncLocalMembershipCommand;
+use Database\Command\MailingListSyncCommand;
 use Database\Command\MailmanFetchListsCommand;
-use Database\Command\MailmanSyncMembershipCommand;
 use Database\Form\Abolish as AbolishForm;
 use Database\Form\Address as AddressForm;
 use Database\Form\Annulment as AnnulmentForm;
@@ -183,9 +179,7 @@ class Module
                 MailmanFetchListsCommand::class => MailmanFetchListsCommandFactory::class,
                 ListmonkFetchListsCommand::class => ListmonkFetchListsCommandFactory::class,
                 MailingListMaintenanceCommand::class => MailingListMaintenanceCommandFactory::class,
-                MailingListSyncLocalMembershipCommand::class => MailingListSyncLocalMembershipCommandFactory::class,
-                MailmanSyncMembershipCommand::class => MailmanSyncMembershipCommandFactory::class,
-                ListmonkSyncMembershipCommand::class => ListmonkSyncMembershipCommandFactory::class,
+                MailingListSyncCommand::class => MailingListSyncCommandFactory::class,
                 GenerateAuthenticationKeysCommand::class => GenerateAuthenticationKeysCommandFactory::class,
                 ApiService::class => ApiServiceFactory::class,
                 FrontPageService::class => FrontPageServiceFactory::class,
